refactor: create SceneRenderer class and use for rendering of Scene
